Move most things out of utils
